+++ reconcile/gql_definitions/terraform_cloudflare_resources/terraform_cloudflare_resources.gql
@@ -1,6 +1,6 @@
 # qenerate: plugin=pydantic_v1
 
-query TerraformResourcesCloudflare {
+query TerraformCloudflareResources {
   namespaces: namespaces_v1 {
     name
     externalResources {
