+++ reconcile/test/test_terraform_cloudflare_resources.py
@@ -1,6 +1,6 @@
 from unittest.mock import patch
 import pytest
-from reconcile.gql_definitions.terraform_resources_cloudflare.terraform_resources_cloudflare import (
+from reconcile.gql_definitions.terraform_cloudflare_resources.terraform_cloudflare_resources import (
     AWSAccountV1,
     CloudflareAccountV1,
     CloudflareZoneRecordV1,
@@ -11,17 +11,17 @@ from reconcile.gql_definitions.terraform_resources_cloudflare.terraform_resource
     NamespaceTerraformProviderResourceCloudflareV1,
     NamespaceTerraformResourceCloudflareZoneV1,
     NamespaceV1,
-    TerraformResourcesCloudflareQueryData,
+    TerraformCloudflareResourcesQueryData,
     TerraformStateAWSV1,
     VaultSecretV1,
 )
-from reconcile.terraform_resources_cloudflare import build_specs
+from reconcile.terraform_cloudflare_resources import build_specs
 from reconcile.utils.external_resource_spec import ExternalResourceSpec
 
 
 @pytest.fixture
 def query_data(external_resources):
-    return TerraformResourcesCloudflareQueryData(
+    return TerraformCloudflareResourcesQueryData(
         namespaces=[
             NamespaceV1(name="namespace1", externalResources=[external_resources])
         ],
@@ -110,7 +110,7 @@ def external_resources(provisioner_config):
 
 def test_build_specs(query_data):
     with patch(
-        "reconcile.terraform_resources_cloudflare.get_github_file", return_value="foo"
+        "reconcile.terraform_cloudflare_resources.get_github_file", return_value="foo"
     ):
         actual = build_specs(query_data)
     expected = [
